backend/requirements.txt
gunicorn==20.1.0
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.2.0
Django==2.2.16
django-cleanup==6.0.0
django-filter==2.4.0
django-jquery==3.1.0
django-minio-storage==0.3.10
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flake8==3.9.2
idna==3.2
importlib-metadata==4.6.3
iniconfig==1.1.1
mccabe==0.6.1
minio==6.0.2
packaging==21.0
Pillow==9.1.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0